Changed the risk impact and probability to be between 0 and 1 Sandra, Krasen, Eemil, Chris

+++ src/main/java/se/turingturtles/ProjectManagement.java
@@ -28,7 +28,7 @@ public interface ProjectManagement {
     List<Task> retrieveTasks();
     void completeTask(Task task);
     int timeSpentOnProject();
-    void createRisk(String name, int impact, int probability);
+    void createRisk(String name, double impact, double probability);
     List<Risk> getProjectRisks();
